+++ packages/frontend/src/components/Button/index.tsx
@@ -6,7 +6,6 @@ import styles from './style.module.scss'
 export type ButtonProps = React.ButtonHTMLAttributes<HTMLButtonElement> & {
   active?: boolean
   styling?: 'primary' | 'secondary' | 'danger'
-  dataTestid?: string
 }
 
 export default function Button({